Update Other Players list in lobby when players leave or join
This adds a function to update the value of the Other Players list in the
lobby. It is called when a user connects and disconnects.

This also removes double entries of `business-team.js` from the define() call in
`main.js`. It also adds a return statement at the end of `business-team.js`
that serves as a module for the file. At the moment it only contains
`updateLobbyList`.